requirements.txt
python-telegram-bot==21.3
aiofiles==23.2.1
aiohttp==3.9.5